+++ resp.h
@@ -9,6 +9,6 @@
 
 enum status resp_dir(int, const char *, const struct request *);
 enum status resp_file(int, const char *, const struct request *,
-                      const struct stat *, const char *, off_t, off_t);
+                      const struct stat *, const char *, size_t, size_t);
 
 #endif /* RESP_H */
